+++ firebase-crashlytics/src/main/java/com/google/firebase/crashlytics/FirebaseCrashlytics.java
@@ -34,7 +34,7 @@ import com.google.firebase.crashlytics.internal.common.CrashlyticsAppQualitySess
 import com.google.firebase.crashlytics.internal.common.CrashlyticsCore;
 import com.google.firebase.crashlytics.internal.common.DataCollectionArbiter;
 import com.google.firebase.crashlytics.internal.common.IdManager;
-import com.google.firebase.crashlytics.internal.concurrency.CrashlyticsWorker;
+import com.google.firebase.crashlytics.internal.concurrency.CrashlyticsWorkers;
 import com.google.firebase.crashlytics.internal.network.HttpRequestFactory;
 import com.google.firebase.crashlytics.internal.persistence.FileStore;
 import com.google.firebase.crashlytics.internal.settings.SettingsController;
@@ -78,6 +78,9 @@ public class FirebaseCrashlytics {
                 + " for "
                 + appIdentifier);
 
+    CrashlyticsWorkers crashlyticsWorkers =
+        new CrashlyticsWorkers(backgroundExecutorService, blockingExecutorService);
+
     FileStore fileStore = new FileStore(context);
     final DataCollectionArbiter arbiter = new DataCollectionArbiter(app);
     final IdManager idManager =
@@ -96,9 +99,6 @@ public class FirebaseCrashlytics {
     RemoteConfigDeferredProxy remoteConfigDeferredProxy =
         new RemoteConfigDeferredProxy(remoteConfigInteropDeferred);
 
-    CrashlyticsWorker commonWorker = new CrashlyticsWorker(backgroundExecutorService);
-    CrashlyticsWorker diskWriteWorker = new CrashlyticsWorker(backgroundExecutorService);
-
     final CrashlyticsCore core =
         new CrashlyticsCore(
             app,
@@ -110,8 +110,7 @@ public class FirebaseCrashlytics {
             fileStore,
             sessionsSubscriber,
             remoteConfigDeferredProxy,
-            commonWorker,
-            diskWriteWorker);
+            crashlyticsWorkers);
 
     final String googleAppId = app.getOptions().getApplicationId();
     final String mappingFileId = CommonUtils.getMappingFileId(context);
@@ -160,7 +159,7 @@ public class FirebaseCrashlytics {
 
     // Kick off actually fetching the settings.
     settingsController
-        .loadSettingsData(commonWorker, blockingExecutorService)
+        .loadSettingsData(crashlyticsWorkers)
         .addOnFailureListener(ex -> Logger.getLogger().e("Error fetching settings.", ex));
 
     final boolean finishCoreInBackground = core.onPreExecute(appData, settingsController);
@@ -437,14 +436,16 @@ public class FirebaseCrashlytics {
   }
 
   /**
-   * Indicates whether or not automatic data collection is enabled
+   * Indicates whether or not automatic data collection is enabled.
    *
    * @return In order of priority:
-   *     <p>If {@link #setCrashlyticsCollectionEnabled(boolean)} is called with a value, use it
-   *     <p>If the <b>firebase_crashlytics_collection_enabled</b> key is in your app’s
-   *     AndroidManifest.xml, use it
-   *     <p>Otherwise, use the default {@link FirebaseApp#isDataCollectionDefaultEnabled()} in
-   *     FirebaseApp
+   *     <ul>
+   *       <li>If {@link #setCrashlyticsCollectionEnabled(boolean)} is called with a value, use it.
+   *       <li>If the <b>firebase_crashlytics_collection_enabled</b> key is in your app’s
+   *           AndroidManifest.xml, use it.
+   *       <li>Otherwise, use the default {@link FirebaseApp#isDataCollectionDefaultEnabled()} in
+   *           FirebaseApp.
+   *     </ul>
    */
   public boolean isCrashlyticsCollectionEnabled() {
     return core.isCrashlyticsCollectionEnabled();