+++ firebase-crashlytics/CHANGELOG.md
@@ -1,11 +1,27 @@
 # Unreleased
-* [feature] Added the `isCrashlyticsCollectionEnabled` API to check if Crashlytics collection is enabled.
-  ([Github #5919](//github.com/firebase/firebase-android-sdk/issues/5919))
-* [fixed] Ensure that on-demand fatal events are never processed on the main thread.
+* [fixed] Improved data consistency for rapid user actions.
+* [fixed] Fixed exception propagation in the case of no default uncaught exception handler.
+* [changed] Internal changes to improve startup time.
+* [changed] Internal changes to the way background tasks are scheduled.
+* [changed] Migrated SDK to use standard Firebase executors.
 
+# 19.1.0
+* [feature] Added the `isCrashlyticsCollectionEnabled` API to check if Crashlytics collection is
+  enabled.
+  (GitHub [#5919](https://github.com/firebase/firebase-android-sdk/issues/5919){: .external})
+* [fixed] Ensure that on-demand fatal events are never processed on the main thread.
+  (GitHub [#4345](https://github.com/firebase/firebase-android-sdk/issues/4345){: .external})
+* [changed] Internal changes to the way session IDs are generated.
+
+
+## Kotlin
+The Kotlin extensions library transitively includes the updated
+`firebase-crashlytics` library. The Kotlin extensions library has no additional
+updates.
 
 # 19.0.3
 * [changed] Update the internal file system to handle long file names.
+
 
 ## Kotlin
 The Kotlin extensions library transitively includes the updated