+++ firebase-vertexai/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Unreleased
+
+
+# 16.0.0-beta05
+* [changed] Merged core networking code into VertexAI from a separate library
 * [feature] added support for `responseSchema` in `GenerationConfig`.
 
 # 16.0.0-beta03